:recycle: rename to home_page
