+++ lib/router/router.dart
@@ -1,5 +1,5 @@
 import 'package:Pilll/domain/initial_setting/initial_setting.dart';
-import 'package:Pilll/domain/home/home.dart';
+import 'package:Pilll/domain/home/home_page.dart';
 import 'package:Pilll/domain/root/root.dart';
 import 'package:Pilll/util/shared_preference/keys.dart';
 import 'package:flutter/material.dart';
